feat: #165 코드리뷰 반영
- 도메인에 common 관련을 login으로 합침
