+++ data/src/main/java/com/mashup/chinchin/data/di/RepositoryModule.kt
@@ -2,14 +2,12 @@ package com.mashup.chinchin.data.di
 
 import com.mashup.chinchin.data.repository.FriendRepositoryImpl
 import com.mashup.chinchin.data.repository.AlarmRepositoryImpl
-import com.mashup.chinchin.data.repository.CommonRepositoryImpl
 import com.mashup.chinchin.data.repository.GroupRepositoryImpl
 import com.mashup.chinchin.data.repository.LoginRepositoryImpl
 import com.mashup.chinchin.data.repository.QuestionnaireRepositoryImpl
 import com.mashup.chinchin.domain.repository.FriendRepository
 import com.mashup.chinchin.domain.repository.AlarmRepository
 import com.mashup.chinchin.data.repository.RecommendedFriendRepositoryImpl
-import com.mashup.chinchin.domain.repository.CommonRepository
 import com.mashup.chinchin.domain.repository.GroupRepository
 import com.mashup.chinchin.domain.repository.LoginRepository
 import com.mashup.chinchin.domain.repository.RecommendedFriendRepository
@@ -48,8 +46,4 @@ interface RepositoryModule {
     fun bindRecommendedFriendRepository(
         recommendedFriendRepositoryImpl: RecommendedFriendRepositoryImpl,
     ): RecommendedFriendRepository
-
-    @Binds
-    @Singleton
-    fun bindCommonRepository(commonRepositoryImpl: CommonRepositoryImpl): CommonRepository
 }
