--- domain/src/main/java/com/mashup/chinchin/domain/repository/CommonRepository.kt
@@ -1,6 +1,0 @@
-package com.mashup.chinchin.domain.repository
-
-interface CommonRepository {
-    fun setIsFirstEnter(isFirstEnter: Boolean)
-    fun getIsFirstEnter(): Boolean
-}
