+++ domain/src/main/java/com/mashup/chinchin/domain/usecase/IsShowCongratulationDialogUseCase.kt
@@ -1,10 +1,10 @@
 package com.mashup.chinchin.domain.usecase
 
-import com.mashup.chinchin.domain.repository.CommonRepository
+import com.mashup.chinchin.domain.repository.LoginRepository
 import javax.inject.Inject
 
 class IsShowCongratulationDialogUseCase @Inject constructor(
-    private val commonRepository: CommonRepository,
+    private val loginRepository: LoginRepository,
 ) {
-    operator fun invoke(): Boolean = commonRepository.getIsFirstEnter()
+    operator fun invoke(): Boolean = loginRepository.getIsFirstEnter()
 }
